session: Fix closing sessions and most memory leaks
It no longer crash when quiting or closing sessions. This is a
step in the right direction to fix all the remaining bugs. Now,
the application can undo everything it does. The main remaining
issues are due to the SIGROK device model. There is no point in
fixing this as the code needs to be extracted to avoid GPL
violation.