quality: Use a better node metadata system
No more editing entries in multiple files.

The metadata are both static and polymorphic. A macro was used
after a couple of failed attempts to use constexpr variadic
templates.

The main limitation is the inability to use string literals in
the template parameters. The workaround was creating a static
function and passing its pointer, but this was ugly.